Add files via upload
when the company that makes rubber bands for skydiving goes out of business and the other company that picks them up produces a subpar replacement, you go for the other option, tube stows.  they are expensive though, but easy to make your own and a lot cheaper.  learn how with this.  i call them tim stows.